// frontend-angular/src/frontend-client/model/showOrderResponse.ts
/**
 * Restaurant Import API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {ParticipantsOrderEntry} from './participantsOrderEntry';
import {SideDish} from './sideDish';
import {ShowOrderDto} from './showOrderDto';
import {DishDto} from './dishDto';


export interface ShowOrderResponse {
    order: ShowOrderDto;
    orderEntries: Array<ParticipantsOrderEntry>;
    currentUserId: string;
    allDishesInRestaurant: Array<DishDto>;
    allDishesByCategory: { [key: string]: Array<DishDto>; };
    dishIdToSideDishesMap: { [key: string]: Array<SideDish>; };
    baseOrderPrice: number;
    totalOrderPrice: number;
}